+++ Packages/com.unity.ugui/Runtime/TMP/TMP_HarfBuzzNative.cs
@@ -86,7 +86,7 @@ namespace TMPro
 
     internal static class TMP_HarfBuzzNative
     {
-#if UNITY_IOS && !UNITY_EDITOR
+#if (UNITY_IOS || UNITY_WEBGL) && !UNITY_EDITOR
         private const string DllName = "__Internal";
 #else
         private const string DllName = "HindiHarfBuzz";